+++ lang/DE/FoodSharePoint.lang.php
@@ -19,6 +19,7 @@ $g_lang['food_share_point_add_success'] = 'Fair-Teiler wurde erfolgreich eingetr
 $g_lang['food_share_point_prepare_success'] = 'Fair-Teiler wurde vorgeschlagen. Ein Botschafter muss jetzt noch Deinen Vorschlag bestätigen.';
 $g_lang['write_teaser'] = 'Nachricht schreiben...';
 $g_lang['attach_image'] = 'Bild anhängen';
+$g_lang['contact_fsp'] = 'Ansprechpartner';
 $g_lang['follower'] = 'Follower';
 $g_lang['bezirk_id'] = 'In welchem Bezirk ist der Fair-Teiler?';
 $g_lang['your_food_share_point'] = 'Fair-Teiler in Deinen Bezirken';
@@ -29,5 +30,5 @@ $g_lang['infotype_desc'] = 'Wie möchtest Du über Updates wie Essenslieferungen
 $g_lang['infotype_email'] = 'über eine E-Mail Benachrichtigung';
 $g_lang['infotype_bell'] = 'nur über interne Infomeldungen';
 $g_lang['desc_desc'] = 'Schreibe hier ein paar grundsätzliche Infos über den Fair-Teiler! Insbesondere wann er zugänglich/geöffnet ist. Du kannst zur Formatierung <a href="http://markdown.de/">Markdown</a> verwenden.';
-$g_lang['bfoodsaver'] = 'Verantwortliche Foodsaver';
+$g_lang['bfoodsaver'] = 'Foodsaver, die Ansprechpartner für den Fair-teiler sind';
 $g_lang['position_search_infobox'] = 'Bitte gib die Adresse des Fair-Teilers im Textfeld zwischen dieser Nachricht und der Karte ein und bestätige diese dann durch Auswahl in der Karte.';